api/chatbot/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
qdrant-client==1.12.1
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
PyMuPDF==1.24.14
markdownify==0.13.1
pypdf==5.1.0
torch==2.5.1
transformers==4.46.3
sentence-transformers==3.3.0
datasets==3.1.0
accelerate==1.1.1
scipy==1.15.1